add reports view for patients
