+++ src/components/reports/doctor.tsx
@@ -7,6 +7,8 @@ import { dateToString } from "../../lib/text";
 import type { VisitReportSummary } from "../../server/api/routers/visitReport";
 import Alert from "../Alert";
 
+import { ReportLine } from "./common";
+
 /**
  * Button for linking to create report form
  */
@@ -18,63 +20,6 @@ const CreateReportButton = () => {
     >
       Create Report +
     </Link>
-  );
-};
-
-type DeleteCallback = ({ id }: { id: string }) => void;
-
-/**
- * Single report summary line
- */
-const ReportLine = ({
-  report,
-  deleteReport,
-}: {
-  report: VisitReportSummary;
-  deleteReport: DeleteCallback;
-}) => {
-  const shortName =
-    report.patient_name?.split(" ").map((value) => {
-      return value.at(0)?.toUpperCase();
-    }) ?? "";
-
-  /**
-   * Deletes the row item
-   */
-  const onClick = () => {
-    deleteReport({ id: report.id });
-  };
-
-  return (
-    <div className="flex flex-row justify-between rounded-xl border border-neutral-200 bg-neutral-100 py-2 px-4 shadow-lg">
-      <div className="flex flex-row space-x-4">
-        <div className="grid h-16 w-16 place-items-center rounded-full border border border-neutral-400 bg-green-300">
-          <span className="text-2xl font-bold">{shortName}</span>
-        </div>
-        <div className="my-auto h-fit space-x-2">
-          <h1 className="inline-block text-xl font-semibold">
-            {report.patient_name}
-          </h1>
-          <span className="inline-block text-lg italic text-neutral-400">
-            on {dateToString(report.date)}
-          </span>
-        </div>
-      </div>
-      <div className="my-auto flex h-fit flex-row space-x-4">
-        <Link
-          href={`/report/${report.id}`}
-          className="group my-auto grid h-12 place-items-center rounded-lg border border-blue-300 bg-blue-200 px-4 font-semibold text-neutral-600 hover:bg-blue-400 hover:text-white"
-        >
-          View Report
-        </Link>
-        <button
-          onClick={onClick}
-          className="group my-auto grid h-12 w-12 place-items-center rounded-lg border border-red-300 bg-red-200 hover:bg-red-400"
-        >
-          <TrashIcon className="w-8 text-neutral-600 group-hover:text-black" />
-        </button>
-      </div>
-    </div>
   );
 };
